Do not literally compare referenced_type of class fields
referenced_type is a mangled name and an optional suffix that
distinguishes multiple definitions. header-abi-linker does not always
generate deterministic referenced_type for multiple definitions and
anonymous types. When header-abi-diff checks class, struct, or union
fields, it should not literally compare the referenced_types. It looks
up the TypeIRs and compares the attributes.

Bug: 355612961
Test: ANDROID_BUILD_TOP=`realpath .` \
      PATH=$PATH:`realpath out/soong/dist/bin` \
      development/vndk/tools/header-checker/tests/test.py
Change-Id: Ie32167f8b62bf587e5d3274451c99fb40d0741c0