Revert "remerge3: Redo 3-way merge for files with conflict markers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/717fbfb7-5177-4b1a-9a2c-72037b2b6d9c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I7530952a510b2ad6ea346b07a54277440dcf6be0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com