Revert "remerge3: Redo 3-way merge for files with conflict markers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/8adb49e7-ead8-4def-ac25-806ebd984a29). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie2ab5dd1c9e50f06b9256aa447c254967c9259bf
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com