Revert "remerge3: Redo 3-way merge for files with conflict markers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a9b65197-820f-490c-8434-5e81ef77342c). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie039a0c301dee7ff4a8158ccd03d333347f7bf08
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com