Revert "remerge3: Redo 3-way merge for files with conflict markers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/90ac4abb-ff9a-491e-9d04-ff61161e540b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic5a8dc318c54e6c5bc19e488bdeb587f7273098b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com