Revert "remerge3: Redo 3-way merge for files with conflict markers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/69b9f87b-4484-4784-a186-cc35c7d3690b). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I1e3f726b90d24b857b5408ee92908b2b680d8146
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com