Revert "remerge3: Redo 3-way merge for files with conflict markers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/94330cc7-02b6-4537-8545-a64b750dc51f). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie0bad52e755d573c2ec6481b7fda049bda4c4252
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com