Revert "remerge3: Redo 3-way merge for files with conflict markers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/d5d3febf-7d21-44e6-b3de-1daa767a4746). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2513d79a05fff500195e54e5c176c9fd9927e357
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com