Revert "remerge3: Redo 3-way merge for files with conflict markers"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/611d018f-1273-475c-a14e-c4ee52aebeb9). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I703cdfaafda5291fe91e0d27ad05f539000d33b9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com