Revert "Don't check in Cargo.lock."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3b2239ac-ca00-4441-bd17-dd955816b94e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I129bb6b94190f51707ff0d3a863fd3ec375148f0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com