Add Android.bps for google_metadata and rooted_path.
So we can run tests at presubmit, with the Android build system.

Requires adding documentation for public methods to satisfy the build
system, so do that as well.

workspace_excludes seems to not work, so exclude all the modules for the
other crates so as not to generate Android.bps for them.

Change-Id: Ie9a1c2101e3e9a30f791a141bda8cd965ef35ccb
Bug: 339424309
Test: treehugger