Use INVALID_DISPLAY id for Trackball event
As we know, trackball events are dispatched based on the focused
window. In a multi-display environment, if display #0 is not focused,
and we still specify the target display as display #0, it may lead to
unexpected ANR. Here, we will default the displayId of trackball events
to -1, allowing the InputDispatcher to determine the specific target
display for dispatching.

Bug: 368378837
Test: Build

Change-Id: I60c8f615e07a1874b15760081ff178376b79e6ec
Signed-off-by: Linnan Li <lilinnan@xiaomi.corp-partner.google.com>