Revert "[MTE] teach stack to handle MTE stack reports"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/023e28ed-844c-45c8-bdd9-ed814d45bfdd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ica3dc84d5f7f8a68bc6dfed8a4903937d491e859
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com