Revert "Move GoogleMetadata to a separate crate."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/7ef324dd-905e-425b-94d6-dfbc16bc7106). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I20b9321987ba058d4e467bd0c54d4a49e96e2322
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com