Revert "Fix NullPointerException in MonkeyMotionEvent"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cdb95a80-3d60-4e99-8465-b54cf058ae16). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I39dc7d27955e027d4919fd5a582a423c45ad565a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com