Revert "Add new command to initialize a new managed repo."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/cdb95a80-3d60-4e99-8465-b54cf058ae16). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie3f7cf8bc985f0f09ffcc08f39268f3cb3b2cea6
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com