Revert "Allow multiple entries for license_text."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/22dd4d6f-ec7d-467f-b8f3-b61e787cd46a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I24965841c58d3a998071081a76c8eae3fb8f26cc
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com