Revert "Allow multiple licenses."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/22dd4d6f-ec7d-467f-b8f3-b61e787cd46a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2aefbfdaaf69b03b6fa7a050680cc0fb387eb70b
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com